requirements.txt
Django==4.1.1
djangorestframework==3.13.1
requests==2.28.1
